requirements.txt
asyncpg
fastapi
uvicorn[standard]
google-cloud-aiplatform
mem0ai
pydantic
python-dotenv
pydantic-settings
vertexai
langchain
python-multipart
langchain-core
langchain-community
langchain-google-genai
httpx
sse-starlette
passlib
python-jose
pydantic[email]
bcrypt
langchain_aws
redis
logger